Final deployment & Rest api/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
boto3==1.16.59
botocore==1.19.59
Brotli==1.0.9
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
grpcio==1.33.1
grpcio-tools==1.33.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
lxml==4.6.2
msgpack==0.6.2
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.3
paho-mqtt==1.5.1
pep517==0.8.2
Pillow==8.1.0
progress==1.5
protobuf==3.13.0
PyJWT==2.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
qrcode==6.1
requests==2.25.1
retrying==1.3.3
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
whitenoise==5.2.0
